Addresing PR comments.
